Extract SlackUser to models package
